requirements.txt
Flask==1.1.2
SQLAlchemy==1.4.46
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.4.4
Flask-CORS==4.0.0
Flask-Migrate==3.1.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
bcrypt==4.0.1
marshmallow==3.19.0
werkzeug==2.0.3
gunicorn==20.1.0
requests==2.28.2 